refactor: Move base model to root and rename modules to models
