+++ src/mlcast/config/base.py
@@ -11,8 +11,8 @@ from pytorch_lightning.loggers import TensorBoardLogger
 
 from ..data.source_data_module import SourceDataDataModule
 from ..data.source_datasets import SourceDataPrecomputedSamplingDataset
-from ..models.base import NowcastLightningModule
-from ..modules.convgru_modules import ConvGruModel
+from ..models.convgru_modules import ConvGruModel
+from ..nowcasting_module import NowcastLightningModule
 
 
 @dataclass
