+++ src/mlcast/config/demo.py
@@ -8,8 +8,8 @@ from torch.utils.data import Dataset as TorchDataset
 
 from ..data.source_data_module import SourceDataDataModule
 from ..data.source_datasets import SourceDataPrecomputedSamplingDataset
-from ..models.base import NowcastLightningModule
-from ..modules.convgru_modules import ConvGruModel
+from ..models.convgru_modules import ConvGruModel
+from ..nowcasting_module import NowcastLightningModule
 
 
 @dataclass
