--- src/mlcast/modules/__init__.py
@@ -1,3 +1,0 @@
-from .convgru_modules import ConvGruModel
-
-__all__ = ["ConvGruModel"]
